requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Talisman==1.1.0
psycopg2-binary==2.9.9
Werkzeug==3.0.1
plotly==5.18.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
